requirements.txt
line-bot-sdk
urllib3
google-cloud-translate